+++ __main__.py
@@ -1,3 +1,3 @@
 import main
 
-main.init()
+main.main()
