+++ docker/dockerfiles/fedora/build-deps/Dockerfile
@@ -1,7 +1,7 @@
 FROM fedora:25
 
 # Misc packages
-RUN dnf -y install pwgen tar psmisc procps findutils iputils net-tools wget logrotate zip findutils git
+RUN dnf -y install pwgen tar psmisc procps findutils iputils net-tools wget logrotate zip findutils git python-pip
 
 # Dependencies for LalkaChat
 RUN dnf -y install wxPython
