+++ modules/helper/system.py
@@ -18,6 +18,8 @@ TRANSLATION_FOLDER = os.path.join(PYTHON_FOLDER, "translations")
 CONF_FOLDER = os.path.join(PYTHON_FOLDER, "conf")
 MODULE_FOLDER = os.path.join(PYTHON_FOLDER, "modules")
 MAIN_CONF_FILE = os.path.join(CONF_FOLDER, "config.cfg")
+HTTP_FOLDER = os.path.join(PYTHON_FOLDER, "http")
+
 GUI_TAG = 'gui'
 WINDOWS = True if sys.platform == 'win32' else False
