+++ modules/messaging/webchat.py
@@ -23,14 +23,13 @@ from modules.helper.html_template import HTML_TEMPLATE
 from modules.helper.message import TextMessage, CommandMessage, SystemMessage, RemoveMessageByIDs
 from modules.helper.module import MessagingModule
 from modules.helper.parser import save_settings, convert_to_dict, update, lc_replace
-from modules.helper.system import THREADS, PYTHON_FOLDER, CONF_FOLDER, EMOTE_FORMAT
+from modules.helper.system import THREADS, PYTHON_FOLDER, CONF_FOLDER, EMOTE_FORMAT, HTTP_FOLDER
 from modules.interface.types import *
 
 logging.getLogger('ws4py').setLevel(logging.ERROR)
 DEFAULT_STYLE = 'default'
 DEFAULT_GUI_STYLE = 'default_gui'
 HISTORY_SIZE = 50
-HTTP_FOLDER = os.path.join(PYTHON_FOLDER, "http")
 s_queue = Queue.Queue()
 log = logging.getLogger('webchat')
 REMOVED_TRIGGER = '%%REMOVED%%'
