+++ src/jenkins/lint_tests/helpers/css_test.py
@@ -21,23 +21,25 @@ def render_scss_file(filename, namespace):
         compiler = Compiler(namespace=namespace, output_style='compressed')
         render_file.write(compiler.compile_string(css_content))
 
-for folder in os.listdir('http'):
-    # Load keys
-    cur_path = os.path.join('http', folder)
-    with open(os.path.join(cur_path, 'settings.json'), 'r') as json_settings:
-        setting_keys = json.load(json_settings)
 
-    css_namespace = Namespace()
-    for key, value in setting_keys.items():
-        for base_class, scss_class in SCSS_MAP.items():
-            if isinstance(value, base_class):
-                css_namespace.set_variable('${}'.format(key), scss_class(value))
-                break
+if __name__ == '__main__':
+    for folder in os.listdir('http'):
+        # Load keys
+        cur_path = os.path.join('http', folder)
+        with open(os.path.join(cur_path, 'settings.json'), 'r') as json_settings:
+            setting_keys = json.load(json_settings)
 
-    for item in os.listdir(os.path.join(cur_path, 'css')):
-        if item.endswith('.css'):
-            pass
-        elif item.endswith('.scss'):
-            render_scss_file(os.path.join(cur_path, 'css', item), css_namespace)
-        else:
-            pass
+        css_namespace = Namespace()
+        for key, value in setting_keys.items():
+            for base_class, scss_class in SCSS_MAP.items():
+                if isinstance(value, base_class):
+                    css_namespace.set_variable('${}'.format(key), scss_class(value))
+                    break
+
+        for item in os.listdir(os.path.join(cur_path, 'css')):
+            if item.endswith('.css'):
+                pass
+            elif item.endswith('.scss'):
+                render_scss_file(os.path.join(cur_path, 'css', item), css_namespace)
+            else:
+                pass
