+++ src/jenkins/root/main.spec
@@ -8,6 +8,7 @@ block_cipher = None
 data = [
     ('modules', 'modules'),
     ('img', 'img'),
+    ('scripts', 'scripts'),
     ('translations', 'translations'),
     ('libs/windows/locales', 'locales'),
     ('libs/windows/icudt.dll', ''),
@@ -23,6 +24,9 @@ a = Analysis(['main.py'],
                             'requests',
                             'irc',
                             'ConfigParser',
+                            'modules.main',
+                            'modules.gui',
+                            'modules.message_handler',
                             'modules.chat.goodgame',
                             'modules.chat.sc2tv',
                             'modules.chat.twitch',
@@ -50,7 +54,7 @@ exe = EXE(pyz,
           debug=False,
           strip=False,
           upx=True,
-          console=True )
+          console=False )
 coll = COLLECT(exe,
                a.binaries,
                a.zipfiles,
